azure/serverless-bigdata-pipeline/functions/requirements.txt
tweet-preprocessor==0.5.0
